+++ src/responses/operations/managebuyofferoperationresponse.h
@@ -14,7 +14,7 @@ class Asset;
 class ManageBuyOfferOperationResponse : public OperationResponse
 {
     Q_OBJECT
-    Q_PROPERTY(qint32 offer_id MEMBER m_offerID)
+    Q_PROPERTY(qint64 offer_id MEMBER m_offerID)
     Q_PROPERTY(QString amount MEMBER m_amount)
     // Price is not implemented yet in horizon
     Q_PROPERTY(QString price MEMBER m_price)
@@ -27,7 +27,7 @@ class ManageBuyOfferOperationResponse : public OperationResponse
     Q_PROPERTY(QString selling_asset_code READ sellingAssetCode WRITE setSellingAssetCode)
     Q_PROPERTY(QString selling_asset_issuer READ sellingAssetIssuer WRITE setSellingAssetIssuer)
 
-    qint32 m_offerID;
+    qint64 m_offerID;
     QString m_amount;
     QString m_price;
 
@@ -45,7 +45,7 @@ class ManageBuyOfferOperationResponse : public OperationResponse
 public:
     ManageBuyOfferOperationResponse(QNetworkReply *reply = nullptr);
     virtual ~ManageBuyOfferOperationResponse();
-    qint32 getOfferId() const;
+    qint64 getOfferId() const;
     QString getAmount() const;
     QString getPrice() const;
     Asset* getBuyingAsset();
